Prediction over the same or a different set of regions (or points).

Realizes predictions that can be useful when researchers are interested in predict a variable observed in one political division of a city (or state) on another division of the same region.

predict_spm(x, ...)

# S3 method for sspm_fit
predict_spm(x, .aggregate = TRUE, ...)

# S3 method for mspm_fit
predict_spm(x, ...)

# S3 method for sf
predict_spm(x, spm_obj, .aggregate = TRUE, n_pts, type, ...)

# S3 method for matrix
predict_spm(x, ...)

Arguments

x
a sf object such that its geometris are either points or polygons.

...
additional parameters

.aggregate
logical. Should the predictions be aggregated? In case the input is only a "fit" object, the aggregation is made over the polygons on which the original data was observed. In case the input x is composed by sf POLYGONS, the aggregation is made over this new partition of the study region.

spm_obj
an object of either class sspm_fit or mspm_fit

n_pts
a numeric scalar standing for number of points to form a grid over the whole region to make the predictions

type
character type of grid to be generated. See st_sample.

Value

an object of class spm_pred